feat(infra): add `ktlint` linter to project and refactor code according it
`ktlint` was addeed to gradle build
