+++ server/app/src/main/kotlin/mu/muse/application/muse/RestConfiguration.kt
@@ -6,7 +6,6 @@ import org.springframework.context.annotation.Configuration
 
 @Configuration
 class RestConfiguration {
-
     @Bean
     fun helloEndpoint() = HelloEndpoint()
 }
