+++ server/app/src/main/kotlin/mu/muse/rest/HelloEndpoint.kt
@@ -3,7 +3,6 @@ package mu.muse.rest
 import org.springframework.web.bind.annotation.GetMapping
 
 class HelloEndpoint {
-
     @GetMapping("/hello")
     fun hello(): String {
         return "Hello, World!"
