+++ server/app/src/test/kotlin/mu/muse/AppTest.kt
@@ -5,9 +5,7 @@ import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
 class AppTest {
-
     @Test
     fun contextLoads() {
-
     }
 }
