+++ tools/scripts/runLinter.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+currentDir=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
+rootDir="$currentDir/../../"
+
+(cd "$rootDir/server" && ./gradlew ktlintCheck)
